Pass param of GameObjObserverClass::Custom as uintptr_t

+++ Code/Scripts/Mission00.cpp
@@ -309,7 +309,7 @@ DECLARE_SCRIPT (MTU_Tutorial_Controller, "")
 		}
 	}
 
-	void Custom (GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom (GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		Vector3 powerup_loc = Vector3(-36.504f,75.985f,1.006f);
 
@@ -1286,7 +1286,7 @@ DECLARE_SCRIPT (MTU_Tutorial_Instructor, "")
 		logan_wait_infantry = false;
 	}
 
-	void Custom (GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom (GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		switch (type)
 		{
@@ -3200,7 +3200,7 @@ DECLARE_SCRIPT (MTU_Trigger_Zone, "")
 		}
 	}
 
-	void Custom (GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom (GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == MTU_TYPE_SYDNEY_IS_RESET)
 		{
@@ -3318,7 +3318,7 @@ DECLARE_SCRIPT (MTU_GDI_Soldier, "")
 		}
 	}
 
-	void Custom (GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom (GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == MTU_TYPE_GDI_SOLDIER_PATROL)
 		{
@@ -3358,7 +3358,7 @@ DECLARE_SCRIPT (MTU_Commando, "")
 		sydney_shot = false;
 	}
 
-	void Custom (GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom (GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == MTU_TYPE_STAR)
 		{
@@ -3463,7 +3463,7 @@ DECLARE_SCRIPT (MTU_Commando_Startup, "")
 
 DECLARE_SCRIPT (MTU_PowerUp_Health, "")
 {
-	void Custom (GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom (GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == CUSTOM_EVENT_POWERUP_GRANTED)
 		{
@@ -3484,7 +3484,7 @@ DECLARE_SCRIPT (MTU_PowerUp_Health, "")
 
 DECLARE_SCRIPT (MTU_PowerUp_Armor, "")
 {
-	void Custom (GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom (GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == CUSTOM_EVENT_POWERUP_GRANTED)
 		{
@@ -3622,7 +3622,7 @@ DECLARE_SCRIPT (MTU_Range_Target_Miss_Commando, "")
 
 DECLARE_SCRIPT (MTU_Range_Powerup, "Powerup_ID:int")
 {
-	void Custom (GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom (GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == CUSTOM_EVENT_POWERUP_GRANTED)
 		{
@@ -3670,7 +3670,7 @@ DECLARE_SCRIPT (MTU_GDI_Vehicle, "Vehicle_ID:int")
 		}
 	}
 
-	void Custom (GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom (GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == CUSTOM_EVENT_VEHICLE_ENTERED)
 		{
@@ -3746,7 +3746,7 @@ DECLARE_SCRIPT (MTU_Building_Controller, "Building_ID:int")
 		}
 	}
 
-	void Custom (GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom (GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == MTU_TYPE_BUILDING_DAMAGEABLE)
 		{
@@ -3909,7 +3909,7 @@ DECLARE_SCRIPT (MSK_Controller, "")
 		Commands->Trigger_Spawner(MSK_NOD_ATK_03);
 	}
 
-	void Custom (GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom (GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == MSK_SOLDIER_DEAD)
 		{
