+++ Code/Scripts/Mission03.cpp
@@ -156,7 +156,7 @@ DECLARE_SCRIPT(M03_Objective_Controller, "")  //1100004
 		}
 	}
 	
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == GUNBOAT_KILLED)
 		{
@@ -770,7 +770,7 @@ DECLARE_SCRIPT(RMV_Test_Big_Gun_Turning, "")
 	}
 
 	
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == GUNBOAT_KILLED)
 		{
@@ -981,7 +981,7 @@ DECLARE_SCRIPT(M03_Gunboat_Controller_RMV, "Receive_Type:int, Receive_Param_For_
 		}
 	}
 	
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == CANNON_KILLED)
 		{
@@ -1181,7 +1181,7 @@ DECLARE_SCRIPT(M03_Gunboat_Controller_RMV, "Receive_Type:int, Receive_Param_For_
 		last = 0;
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		char *list[11];
 		
@@ -1296,7 +1296,7 @@ DECLARE_SCRIPT(M03_Outro_Cinematic, "")
 		}
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if ((type == 8000) && (param >= 3))
 		{
@@ -1399,7 +1399,7 @@ DECLARE_SCRIPT(RMV_Volcano_And_Lava_Ball_Creator, "")
 
 	}
 	
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if ((type == 500) && (param == 500))
 		{
@@ -1595,7 +1595,7 @@ DECLARE_SCRIPT(M03_CommCenter_SateliteDish_Controller_JDG, "")
 		}
 	}
 
-	void Custom( GameObject * obj, int type, int param, GameObject * sender ) override
+	void Custom( GameObject * obj, int type, uintptr_t param, GameObject * sender ) override
 	{
 		if (param == stop_sounds)
 		{
@@ -1646,7 +1646,7 @@ DECLARE_SCRIPT(M03_CommCenter_SateliteDish_Controller_JDG, "")
 		}
 	}
 
-	void Custom( GameObject * obj, int type, int param, GameObject * sender ) override
+	void Custom( GameObject * obj, int type, uintptr_t param, GameObject * sender ) override
 	{
 		if (param == stop_sounds)
 		{
@@ -1701,7 +1701,7 @@ DECLARE_SCRIPT(M03_Ambient_Birdcall_Controller_JDG, "")
 		}
 	}
 
-	void Custom( GameObject * obj, int type, int param, GameObject * sender ) override
+	void Custom( GameObject * obj, int type, uintptr_t param, GameObject * sender ) override
 	{
 		if (param == stop_birdcalls)
 		{
@@ -1843,7 +1843,7 @@ DECLARE_SCRIPT(M03_Announce_PowerPlant_Controller_JDG, "")
 		}
 	}
 
-	void Custom( GameObject * obj, int type, int param, GameObject * sender ) override
+	void Custom( GameObject * obj, int type, uintptr_t param, GameObject * sender ) override
 	{
 		
 
@@ -2265,7 +2265,7 @@ DECLARE_SCRIPT(M03_Announce_Refinery_Controller_JDG, "")
 		}
 	}
 
-	void Custom( GameObject * obj, int type, int param, GameObject * sender ) override
+	void Custom( GameObject * obj, int type, uintptr_t param, GameObject * sender ) override
 	{
 		if (param == stop_announcements)
 		{
@@ -2752,7 +2752,7 @@ DECLARE_SCRIPT(M03_Announce_CommCenter_Controller_JDG, "")
 		}
 	}
 
-	void Custom( GameObject * obj, int type, int param, GameObject * sender ) override
+	void Custom( GameObject * obj, int type, uintptr_t param, GameObject * sender ) override
 	{
 		if (param == stop_announcements)
 		{
@@ -3048,7 +3048,7 @@ DECLARE_SCRIPT(M03_Initial_Powerups, "")
 		}	
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if ((type == 12176) && (param == 12176))
 		{
@@ -3082,7 +3082,7 @@ DECLARE_SCRIPT(M03_Commando_Script, "Controller_ID:int")
 		}
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == 3000)
 		{
@@ -3167,7 +3167,7 @@ DECLARE_SCRIPT(M03_Move_Commando_To_Start, "")
 		}
 	}
 	
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if ((type == 100) && (param == 100))
 		{
@@ -3202,7 +3202,7 @@ DECLARE_SCRIPT(M03_Chinook_Drop_Soldiers_GDI, "Controller_ID:int")
 		return type - 4001;
 	}
 	
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		GameObject *soldier;
 		
@@ -3335,7 +3335,7 @@ DECLARE_SCRIPT(M03_Chinook_Reinforcements, "Beach_Preset:string, Inlet_Preset:st
 		}
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == 2000)
 		{
@@ -3573,7 +3573,7 @@ DECLARE_SCRIPT(M03_Chinook_Spawned_Soldier_GDI, "Area:int, Send_Type_When_Killed
 	void Poked(GameObject * obj, GameObject * poker) override
 	{
 		int has_escort = 1;
-		Commands->Send_Custom_Event(obj, poker, 3000, (int)&has_escort, 0.0f);
+		Commands->Send_Custom_Event(obj, poker, 3000, (uintptr_t)&has_escort, 0.0f);
 		if (has_escort == 1)
 		{
 
@@ -3633,7 +3633,7 @@ DECLARE_SCRIPT(M03_Chinook_Spawned_Soldier_GDI, "Area:int, Send_Type_When_Killed
 		}
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == 999 && param == 999)
 		{
@@ -3709,7 +3709,7 @@ DECLARE_SCRIPT(M03_Objective_Tracker, "")
 		number_completed = 0;
 	}
 	
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if ((type == 300) || (type == 308) || (type == 309))
 		{
@@ -3729,7 +3729,7 @@ DECLARE_SCRIPT(M03_Objective_Tracker, "")
 
 DECLARE_SCRIPT(M03_Intro_Substitute, "")
 {
-	void Custom( GameObject * obj, int type, int param, GameObject * sender ) override
+	void Custom( GameObject * obj, int type, uintptr_t param, GameObject * sender ) override
 	{
 		if (type == 999)
 		{
@@ -3833,7 +3833,7 @@ DECLARE_SCRIPT(M03_Alternate_Sam_Site, "Chinook_Controller_ID:int")
 		
 	}
 	
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (Commands->Find_Object(1100003) == NULL)
 		{
@@ -3906,7 +3906,7 @@ DECLARE_SCRIPT(M03_Flyover_Controller, "")
 		comanches = true;
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == 511 && param == 511)
 		{
@@ -3985,7 +3985,7 @@ DECLARE_SCRIPT(M03_Destroyed_Chinook, "Controller_ID:int, Simple_ID:int")
 
 DECLARE_SCRIPT(M03_Chinook_Fodder_Creator, "")
 {
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (Commands->Find_Object(1100003) == NULL)
 		{
@@ -4082,7 +4082,7 @@ DECLARE_SCRIPT(M03_Tailgun_Fodder_Zone, "Spawner_ID_1:int, Spawner_ID_2:int, Spa
 		}
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == 5000)
 		{
@@ -4187,7 +4187,7 @@ DECLARE_SCRIPT(M03_Inlet_Nod_Reinforcements, "")
 		}
 	}
 	
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if ((type == 300) && (param == 300))
 		{
@@ -4345,7 +4345,7 @@ DECLARE_SCRIPT(M03_Base_Harvester, "Tiberium_Loc:vector3, Dock_Location:vector3,
 		Commands->Action_Goto(Owner(), params);
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == 622 && param == 622)
 		{
@@ -4474,7 +4474,7 @@ DECLARE_SCRIPT(M03_Reinforce_Area, "")
 		}
 	}*/
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == 5000)
 		{
@@ -4643,7 +4643,7 @@ DECLARE_SCRIPT(M03_Officer_With_Key_Card2, "")
 
 DECLARE_SCRIPT(M03_Key_Card, "")
 {
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		GameObject * con = Commands->Find_Object(1100004);
 		if (con)
@@ -4834,7 +4834,7 @@ DECLARE_SCRIPT(M03_Chinook_ParaDrop, "Preset:string")
 		}
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == 23000 && param == 23000)
 		{
@@ -4879,7 +4879,7 @@ DECLARE_SCRIPT(M03_Reinforcement_Chinook, "Controller_ID:int")
 
 DECLARE_SCRIPT(M03_Staged_Conversation_1, "Soldier_1_ID:int, Soldier_2_ID:int")
 {
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == 1000 && param == 1000)
 		{
@@ -4945,7 +4945,7 @@ DECLARE_SCRIPT(M03_Staged_Conversation_Soldier, "")
 		Commands->Innate_Soldier_Enable_Bullet_Heard(obj, true);
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		Commands->Set_Innate_Is_Stationary(obj, false);
 		Commands->Innate_Soldier_Enable_Footsteps_Heard(obj, true);
@@ -5078,7 +5078,7 @@ DECLARE_SCRIPT(M03_SAM_Site_Logic, "")
 		Commands->Attach_Script(destroyedSam, "M03_Destroyed_SAM_Site", "");
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == M00_CUSTOM_SAM_SITE_IGNORE && current < 10 && current >= 0)
 		{
@@ -5154,7 +5154,7 @@ DECLARE_SCRIPT(M03_Beach_Scenario_Controller, "")
 		count = 0;
 	}
 	
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		count++;
 		if (count >= 4)
@@ -5193,7 +5193,7 @@ DECLARE_SCRIPT(M03_Beach_Scenario_Controller, "")
 		active = false;
 	}
 	
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == 100 && param == 100)
 		{
@@ -5287,7 +5287,7 @@ DECLARE_SCRIPT(M03_Beach_Scenario_Controller, "")
 
 /*DECLARE_SCRIPT(M03_Dock_Evacuation_Controller, "")
 {
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == 100 && param == 100)
 		{
@@ -5438,7 +5438,7 @@ DECLARE_SCRIPT(DLS_Volcano_Active, "Receive_Type=0:int, Receive_Param=0:int, Vol
 
 	}
 
-	void Custom(GameObject* obj, int type, int param, GameObject* sender) override
+	void Custom(GameObject* obj, int type, uintptr_t param, GameObject* sender) override
 	{
 		if((type == receive_type) && (param == receive_param))
 		{
@@ -5845,14 +5845,14 @@ DECLARE_SCRIPT(M03_Area_Troop_Counter, "")
 		killed_count[0] = killed_count[1] = killed_count[2] = 0;
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == 1000 && param == 1000)
 		{
 			int area = -1;
 			int target_count = -1;
-			Commands->Send_Custom_Event(obj, obj, 5000, (int)&area, 0.0f);
-			Commands->Send_Custom_Event(obj, obj, 6300, (int)&target_count, 0.0f);
+			Commands->Send_Custom_Event(obj, obj, 5000, (uintptr_t)&area, 0.0f);
+			Commands->Send_Custom_Event(obj, obj, 6300, (uintptr_t)&target_count, 0.0f);
 			if (area >= 0 && area <= 2)
 			{
 				area_count[area]--;
@@ -5974,7 +5974,7 @@ DECLARE_SCRIPT (M03_CommCenter_Warning, "")
 		already_entered = false;
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == MCT_ACCESSED)
 		{
@@ -6012,7 +6012,7 @@ DECLARE_SCRIPT (M03_PowerPlant_Warning, "")
 		already_entered = false;
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == MCT_ACCESSED)
 		{
@@ -6058,7 +6058,7 @@ DECLARE_SCRIPT (M03_CommCenter_Arrow, "")
 		}
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (type == MCT_ACCESSED)
 		{
@@ -6147,7 +6147,7 @@ DECLARE_SCRIPT (M03_Comm_Killed, "")
 		max_health = Commands->Get_Health (obj);
 	}
 
-	void Custom (GameObject *obj, int type, int param, GameObject *sender) override
+	void Custom (GameObject *obj, int type, uintptr_t param, GameObject *sender) override
 	{
 		if (type == BASE_ENTERED)
 		{
@@ -6186,7 +6186,7 @@ DECLARE_SCRIPT (M03_Conversation_Zone, "Conv_Num:int")
 		first_conv_played = false;
 	}
 
-	void Custom (GameObject *obj, int type, int param, GameObject *sender) override
+	void Custom (GameObject *obj, int type, uintptr_t param, GameObject *sender) override
 	{
 		if (type == ENTERED)
 		{
@@ -6384,7 +6384,7 @@ DECLARE_SCRIPT(M03_Holograph_EntryZone_JDG, "")
 		}
 	}
 
-	void Custom( GameObject * obj, int type, int param, GameObject * sender ) override
+	void Custom( GameObject * obj, int type, uintptr_t param, GameObject * sender ) override
 	{
 		if (param == 8000)
 		{
@@ -6442,7 +6442,7 @@ DECLARE_SCRIPT(M03_KaneHead_JDG, "") //2017221
 		}
 	}
 
-	void Custom( GameObject * obj, int type, int param, GameObject * sender ) override
+	void Custom( GameObject * obj, int type, uintptr_t param, GameObject * sender ) override
 	{
 		if (param == 8000)//conversation is over--go away
 		{
@@ -6594,7 +6594,7 @@ DECLARE_SCRIPT(M03_Beach_Reinforce, "")
 		}
 	}	
 
-	void Custom (GameObject *obj, int type, int param, GameObject *sender) override
+	void Custom (GameObject *obj, int type, uintptr_t param, GameObject *sender) override
 	{
 		if (type == LOCATION)
 		{
@@ -6689,7 +6689,7 @@ DECLARE_SCRIPT(M03_Engineer_Target, "")
 		full_health = Commands->Get_Health (obj);
 	}
 
-	void Custom (GameObject *obj, int type, int param, GameObject *sender) override
+	void Custom (GameObject *obj, int type, uintptr_t param, GameObject *sender) override
 	{
 		if (type == ENGINEER)
 		{
@@ -6809,7 +6809,7 @@ DECLARE_SCRIPT(M03_Radar_Scramble, "")
 		already_entered = false;				
 	}
 
-	void Custom( GameObject * obj, int type, int param, GameObject * sender ) override
+	void Custom( GameObject * obj, int type, uintptr_t param, GameObject * sender ) override
 	{
 		if (type == ENTERED)
 		{
@@ -6926,7 +6926,7 @@ DECLARE_SCRIPT (M10_Elevator_All_Controller, "")
 		}
 	}
 
-	void Custom (GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom (GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		
 		if (type == ACTIVATE)
@@ -7009,7 +7009,7 @@ DECLARE_SCRIPT(M03_Paratrooper_Run, "")
 		
 	}
 
-	void Custom(GameObject * obj, int type, int param, GameObject * sender) override
+	void Custom(GameObject * obj, int type, uintptr_t param, GameObject * sender) override
 	{
 		if (!initial_damage)
 		{
@@ -7089,7 +7089,7 @@ DECLARE_SCRIPT (M03_Damage_Modifier_All, "Damage_multiplier:float")
 
 DECLARE_SCRIPT(M03_Cine_Explosion, "")
 {
-	void Custom (GameObject *obj, int type, int param, GameObject *sender) override
+	void Custom (GameObject *obj, int type, uintptr_t param, GameObject *sender) override
 	{
 		if (type == 600)
 		{
@@ -7112,7 +7112,7 @@ DECLARE_SCRIPT(M03_ConYardSeen, "")
 		already_entered = false;				
 	}
 
-	void Custom( GameObject * obj, int type, int param, GameObject * sender ) override
+	void Custom( GameObject * obj, int type, uintptr_t param, GameObject * sender ) override
 	{
 		if (type == 900)
 		{
@@ -7132,7 +7132,7 @@ DECLARE_SCRIPT(M03_ConYardSeen, "")
 
 DECLARE_SCRIPT(M03_DataDiscMessage, "")
 {
-	void Custom( GameObject * obj, int type, int param, GameObject * sender ) override
+	void Custom( GameObject * obj, int type, uintptr_t param, GameObject * sender ) override
 	{
 		if ( type == CUSTOM_EVENT_POWERUP_GRANTED ) 
 		{
